# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WishShop API",
        "contact": {
            "email": "lijy91@foxmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "WishShop Server (Local)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "WishShop Server (Local)"
        },
        {
            "url": "https://wishshop-api-dev.thecode.me",
            "description": "WishShop Server (Development)"
        },
        {
            "url": "https://wishshop-api.thecode.me",
            "description": "WishShop Server (Production)"
        }
    ],
    "paths": {
...
        "/stores/{store_id}/brands": {
            "get": {
                "tags": [
                    "brands"
                ],
                "summary": "Get all Brands",
                "operationId": "App\\Http\\Controllers\\BrandController::index",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/brands/{brand_id}": {
            "get": {
                "tags": [
                    "brands"
                ],
                "summary": "Get a specific brand",
                "operationId": "App\\Http\\Controllers\\BrandController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "brand_id",
                        "in": "path",
                        "description": "Brand Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get all Categories",
                "operationId": "App\\Http\\Controllers\\CategoryController::index",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Parent Id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/categories/{category_id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get a specific category",
                "operationId": "App\\Http\\Controllers\\CategoryController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/collections": {
            "get": {
                "tags": [
                    "collections"
                ],
                "summary": "Get all Collections",
                "operationId": "App\\Http\\Controllers\\CollectionController::index",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Parent Id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/collections/{collection_id}": {
            "get": {
                "tags": [
                    "collections"
                ],
                "summary": "Get a specific collection",
                "operationId": "App\\Http\\Controllers\\CollectionController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "collection_id",
                        "in": "path",
                        "description": "Collection Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get all Products",
                "operationId": "App\\Http\\Controllers\\ProductController::index",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/stores/{store_id}/products/{product_id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get a specific product",
                "operationId": "App\\Http\\Controllers\\ProductController::show",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "jwt_auth": {
                "type": "apiKey",
                "description": "To make authenticated requests via http using the built in methods, you will need to set an authorization header as follows:\\n```\\nAuthorization: Bearer {yourtokenhere}\\n```\\n",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}